+++ src/anemoi/inference/runner.py
@@ -145,9 +145,6 @@ class Runner(Context):
         # TODO: Check for user provided forcings
 
         for source in self.constant_forcings_inputs:
-            if source is None:
-                # When the constants are already in the input state
-                continue
             LOG.info("Constant forcings input: %s %s (%s)", source, source.variables, dates)
             arrays = source.load_forcings(input_state, dates)
             for name, forcing in zip(source.variables, arrays):
@@ -164,6 +161,15 @@ class Runner(Context):
                 self._input_kinds[name] = Kind(forcing=True, constant=True, **source.kinds)
 
     def prepare_input_tensor(self, input_state, dtype=np.float32):
+
+        if "latitudes" not in input_state:
+            input_state["latitudes"] = self.checkpoint.latitudes
+
+        if "longitudes" not in input_state:
+            input_state["longitudes"] = self.checkpoint.longitudes
+
+        if input_state.get("latitudes") is None or input_state.get("longitudes") is None:
+            raise ValueError("Input state must contain 'latitudes' and 'longitudes'")
 
         typed_variables = self.checkpoint.typed_variables
 
@@ -353,6 +359,7 @@ class Runner(Context):
         # batch is always 1
 
         for source in self.dynamic_forcings_inputs:
+
             forcings = source.load_forcings(state, [date])  # shape: (variables, dates, values)
 
             forcings = np.squeeze(forcings, axis=1)  # Drop the dates dimension